requirements.txt
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
chromedriver-autoinstaller==0.6.4
colorama==0.4.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
loguru==0.7.3
outcome==1.3.0.post0
packaging==24.2
peewee==3.17.9
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.1.0
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
urllib3==2.3.0
websocket-client==1.8.0
win32_setctime==1.2.0
wsproto==1.2.0
